+++ crates/alkahest-renderer/src/ecs/common.rs
@@ -1,6 +1,7 @@
 use std::fmt::Display;
 
 use ecolor::Color32;
+use glam::Vec3;
 
 /// Tiger entity world ID
 #[derive(Copy, Clone)]
@@ -53,6 +54,7 @@ impl Display for Icon {
 pub struct Label {
     pub label: String,
     pub default: bool,
+    pub offset: Vec3,
 }
 
 impl Display for Label {
@@ -67,7 +69,12 @@ impl Label {
         Self {
             label: s.to_string(),
             default: true,
+            offset: Vec3::new(0.0, 0.0, 0.0),
         }
+    }
+    pub fn with_offset(mut self, x: f32, y: f32, z: f32) -> Self {
+        self.offset = Vec3::new(x, y, z);
+        self
     }
 }
 
@@ -76,6 +83,7 @@ impl From<&str> for Label {
         Self {
             label: s.to_string(),
             default: false,
+            offset: Vec3::new(0.0, 0.0, 0.0),
         }
     }
 }
@@ -85,6 +93,7 @@ impl From<String> for Label {
         Self {
             label: s,
             default: false,
+            offset: Vec3::new(0.0, 0.0, 0.0),
         }
     }
 }