+++ crates/alkahest/src/gui/hotkeys.rs
@@ -1,13 +1,19 @@
 use alkahest_renderer::{
-    camera::{
-        tween::ease_out_exponential,
-        Camera,
+    camera::{tween::ease_out_exponential, Camera},
+    ecs::{
+        common::Hidden,
+        hierarchy::{Children, Parent},
+        resources::SelectedEntity,
     },
-    ecs::{common::Hidden, resources::SelectedEntity},
     renderer::RendererShared,
+    util::scene::SceneExt,
 };
 
-use crate::{maplist::MapList, resources::Resources, util::action::{ActionList, TweenAction}};
+use crate::{
+    maplist::MapList,
+    resources::Resources,
+    util::action::{ActionList, TweenAction},
+};
 
 pub const SHORTCUT_DELETE: egui::KeyboardShortcut =
     egui::KeyboardShortcut::new(egui::Modifiers::SHIFT, egui::Key::Delete);
@@ -35,11 +41,29 @@ pub const SHORTCUT_GAZE: egui::KeyboardShortcut =
 pub const SHORTCUT_MAP_SWAP: egui::KeyboardShortcut =
     egui::KeyboardShortcut::new(egui::Modifiers::NONE, egui::Key::I);
 
-    pub const SHORTCUT_MAP_PREV: egui::KeyboardShortcut =
+pub const SHORTCUT_MAP_PREV: egui::KeyboardShortcut =
     egui::KeyboardShortcut::new(egui::Modifiers::NONE, egui::Key::PageUp);
 
-    pub const SHORTCUT_MAP_NEXT: egui::KeyboardShortcut =
+pub const SHORTCUT_MAP_NEXT: egui::KeyboardShortcut =
     egui::KeyboardShortcut::new(egui::Modifiers::NONE, egui::Key::PageDown);
+
+pub const SHORTCUT_ADD_ROUTE_NODE_NEXT: egui::KeyboardShortcut =
+    egui::KeyboardShortcut::new(egui::Modifiers::NONE, egui::Key::Plus);
+
+pub const SHORTCUT_ADD_ROUTE_NODE_PREV: egui::KeyboardShortcut =
+    egui::KeyboardShortcut::new(egui::Modifiers::NONE, egui::Key::Minus);
+
+pub const SHORTCUT_SELECT_PARENT: egui::KeyboardShortcut =
+    egui::KeyboardShortcut::new(egui::Modifiers::NONE, egui::Key::ArrowUp);
+
+pub const SHORTCUT_SELECT_CHILD: egui::KeyboardShortcut =
+    egui::KeyboardShortcut::new(egui::Modifiers::NONE, egui::Key::ArrowDown);
+
+pub const SHORTCUT_SELECT_NEXT_CHILD: egui::KeyboardShortcut =
+    egui::KeyboardShortcut::new(egui::Modifiers::NONE, egui::Key::ArrowRight);
+
+pub const SHORTCUT_SELECT_PREV_CHILD: egui::KeyboardShortcut =
+    egui::KeyboardShortcut::new(egui::Modifiers::NONE, egui::Key::ArrowLeft);
 
 pub fn process_hotkeys(ctx: &egui::Context, resources: &mut Resources) {
     if ctx.input_mut(|i| i.consume_shortcut(&SHORTCUT_UNHIDE_ALL)) {
@@ -62,15 +86,35 @@ pub fn process_hotkeys(ctx: &egui::Context, resources: &mut Resources) {
     }
 
     if ctx.input_mut(|i| i.consume_shortcut(&SHORTCUT_MAP_PREV)) {
-        resources.get_mut::<MapList>().set_current_map_prev(resources);
+        resources
+            .get_mut::<MapList>()
+            .set_current_map_prev(resources);
     }
 
     if ctx.input_mut(|i| i.consume_shortcut(&SHORTCUT_MAP_NEXT)) {
-        resources.get_mut::<MapList>().set_current_map_next(resources);
+        resources
+            .get_mut::<MapList>()
+            .set_current_map_next(resources);
     }
 
     if ctx.input_mut(|i| i.consume_shortcut(&SHORTCUT_GAZE)) {
         goto_gaze(resources);
+    }
+
+    if ctx.input_mut(|i| i.consume_shortcut(&SHORTCUT_SELECT_PARENT)) {
+        select_parent(resources);
+    }
+
+    if ctx.input_mut(|i| i.consume_shortcut(&SHORTCUT_SELECT_CHILD)) {
+        select_child(resources);
+    }
+
+    if ctx.input_mut(|i| i.consume_shortcut(&SHORTCUT_SELECT_NEXT_CHILD)) {
+        select_child_offset(resources, true);
+    }
+
+    if ctx.input_mut(|i| i.consume_shortcut(&SHORTCUT_SELECT_PREV_CHILD)) {
+        select_child_offset(resources, false);
     }
 }
 
@@ -113,5 +157,51 @@ fn goto_gaze(resources: &mut Resources) {
             None,
             0.7,
         ));
+    }
+}
+
+fn select_parent(resources: &mut Resources) {
+    let mut selected = resources.get_mut::<SelectedEntity>();
+    let mut maps = resources.get_mut::<MapList>();
+    if let Some(current) = selected.selected() {
+        if let Some(map) = maps.current_map_mut() {
+            if let Some(parent) = map.scene.get_parent(current) {
+                selected.select(parent);
+            }
+        }
+    }
+}
+
+fn select_child(resources: &mut Resources) {
+    let mut selected = resources.get_mut::<SelectedEntity>();
+    let mut maps = resources.get_mut::<MapList>();
+    if let Some(current) = selected.selected() {
+        if let Some(map) = maps.current_map_mut() {
+            if let Ok(children) = map.scene.get::<&Children>(current) {
+                selected.select(children.0[0]);
+            }
+        }
+    }
+}
+
+fn select_child_offset(resources: &mut Resources, add: bool) {
+    let mut selected = resources.get_mut::<SelectedEntity>();
+    let mut maps = resources.get_mut::<MapList>();
+    if let Some(current) = selected.selected() {
+        if let Some(map) = maps.current_map_mut() {
+            if let Some(parent) = map.scene.get_parent(current) {
+                if let Ok(children) = map.scene.get::<&Children>(parent) {
+                    let i = children
+                        .0
+                        .iter()
+                        .position(|&ent| ent == current)
+                        .unwrap_or(0);
+                    let new_i = if add { i + 1 } else { i - 1 };
+                    if let Some(next) = children.0.get(new_i) {
+                        selected.select(*next);
+                    }
+                }
+            }
+        }
     }
 }