+++ crates/alkahest/src/gui/inspector/mod.rs
@@ -16,12 +16,12 @@ use alkahest_renderer::{
         resources::SelectedEntity,
         tags::{insert_tag, remove_tag, EntityTag, Tags},
         transform::{OriginalTransform, Transform, TransformFlags},
-        utility::{Beacon, Route, Ruler, Sphere},
+        utility::{Beacon, Route, RouteNode, Ruler, Sphere},
         Scene,
     },
     icons::{
-        ICON_ACCOUNT_CONVERT, ICON_EYE_ARROW_RIGHT_OUTLINE, ICON_EYE_OFF_OUTLINE, ICON_HUMAN_CHILD,
-        ICON_HUMAN_MALE, ICON_HUMAN_MALE_FEMALE_CHILD, ICON_POKEBALL,
+        ICON_ACCOUNT_CONVERT, ICON_EYE_ARROW_RIGHT_OUTLINE, ICON_EYE_OFF_OUTLINE, ICON_HUMAN_MALE,
+        ICON_HUMAN_MALE_FEMALE_CHILD, ICON_POKEBALL,
     },
     renderer::RendererShared,
     shader::shader_ball::ShaderBallComponent,
@@ -123,10 +123,31 @@ pub fn show_inspector_panel(
                 .clicked()
                 || ui.input_mut(|i| i.consume_shortcut(&SHORTCUT_DELETE)))
         {
+            // Remove all children
+            if let Some(children) = e.get::<&Children>() {
+                for child_ent in &children.0 {
+                    cmd.despawn(*child_ent);
+                }
+            }
+            // Remove from parent
+            if let Some(parent) = e.get::<&Parent>() {
+                if let Ok(parent) = scene.entity(parent.0) {
+                    if let Some(mut children) = parent.get::<&mut Children>() {
+                        children.0.retain(|child| *child != e.entity());
+                    }
+                }
+                resources.get_mut::<SelectedEntity>().select(parent.0);
+            }
             cmd.despawn(ent);
         }
 
-        if ui
+        if e.has::<RouteNode>() {
+            ui.label(
+                RichText::new(if visible { ICON_EYE } else { ICON_EYE_OFF })
+                    .size(24.0)
+                    .strong(),
+            );
+        } else if ui
             .button(
                 RichText::new(if visible { ICON_EYE } else { ICON_EYE_OFF })
                     .size(24.0)
@@ -139,6 +160,11 @@ pub fn show_inspector_panel(
                 cmd.insert_one(ent, Hidden);
             } else {
                 cmd.remove_one::<Hidden>(ent);
+            }
+            if e.has::<Route>() {
+                if let Some(route) = e.get::<&Route>() {
+                    route.fixup_visiblity(scene, cmd, ent, Some(visible));
+                }
             }
         }
 
@@ -195,9 +221,9 @@ pub fn show_inspector_panel(
                         |ui| {
                             for c in children.iter() {
                                 let title = if let Ok(label) = scene.get::<&Label>(*c) {
-                                    format!("{label} (id {})", e.entity().id())
+                                    format!("{label} (id {})", c.id())
                                 } else {
-                                    format!("Entity {}", e.entity().id())
+                                    format!("Entity {}", c.id())
                                 };
 
                                 if ui.selectable_label(false, title).clicked() {
@@ -235,7 +261,7 @@ pub fn show_inspector_panel(
         };
         ui.separator();
     }
-    show_inspector_components(ui, scene, e, resources);
+    show_inspector_components(ui, scene, cmd, e, resources);
 
     if global_changed {
         if global {
@@ -249,12 +275,13 @@ pub fn show_inspector_panel(
 fn show_inspector_components(
     ui: &mut egui::Ui,
     scene: &Scene,
+    cmd: &mut hecs::CommandBuffer,
     e: EntityRef<'_>,
     resources: &Resources,
 ) {
     if let Some(mut t) = e.get::<&mut Transform>() {
         inspector_component_frame(ui, "Transform", ICON_AXIS_ARROW, |ui| {
-            t.show_inspector_ui(scene, e, ui, resources);
+            t.show_inspector_ui(scene, cmd, e, ui, resources);
         });
     }
 
@@ -263,7 +290,7 @@ fn show_inspector_components(
 			$(
 				if let Some(mut component) = e.get::<&mut $component>() {
 					inspector_component_frame(ui, <$component>::inspector_name(), <$component>::inspector_icon(), |ui| {
-						component.show_inspector_ui(scene, e, ui, resources);
+						component.show_inspector_ui(scene, cmd, e, ui, resources);
 					});
 				}
 			)*
@@ -276,6 +303,7 @@ fn show_inspector_components(
         Sphere,
         Beacon,
         Route,
+        RouteNode,
         DynamicModelComponent,
         LightRenderer,
         SLightCollection,
@@ -314,6 +342,7 @@ pub(super) trait ComponentPanel {
     fn show_inspector_ui<'s>(
         &mut self,
         _: &'s Scene,
+        _: &mut hecs::CommandBuffer,
         _: EntityRef<'s>,
         _: &mut egui::Ui,
         _: &Resources,
@@ -337,6 +366,7 @@ impl ComponentPanel for Transform {
     fn show_inspector_ui(
         &mut self,
         scene: &Scene,
+        _: &mut hecs::CommandBuffer,
         e: EntityRef<'_>,
         ui: &mut egui::Ui,
         resources: &Resources,
@@ -484,7 +514,14 @@ impl ComponentPanel for DynamicModelComponent {
         true
     }
 
-    fn show_inspector_ui(&mut self, _: &Scene, _: EntityRef<'_>, ui: &mut egui::Ui, _: &Resources) {
+    fn show_inspector_ui(
+        &mut self,
+        _: &Scene,
+        _: &mut hecs::CommandBuffer,
+        _: EntityRef<'_>,
+        ui: &mut egui::Ui,
+        _: &Resources,
+    ) {
         ui.horizontal(|ui| {
             ui.strong("Hash:");
             ui.label(self.model.hash.to_string());
@@ -559,7 +596,14 @@ impl ComponentPanel for ShaderBallComponent {
         true
     }
 
-    fn show_inspector_ui(&mut self, _: &Scene, _: EntityRef<'_>, ui: &mut egui::Ui, _: &Resources) {
+    fn show_inspector_ui(
+        &mut self,
+        _: &Scene,
+        _: &mut hecs::CommandBuffer,
+        _: EntityRef<'_>,
+        ui: &mut egui::Ui,
+        _: &Resources,
+    ) {
         ui.horizontal(|ui| {
             ui.strong("Color:");
             ui.color_edit_button_rgb(self.color.as_mut());
@@ -621,6 +665,7 @@ impl ComponentPanel for NodeMetadata {
     fn show_inspector_ui<'s>(
         &mut self,
         _: &'s Scene,
+        _: &mut hecs::CommandBuffer,
         _: EntityRef<'s>,
         ui: &mut Ui,
         _: &Resources,
@@ -668,6 +713,7 @@ impl ComponentPanel for SRespawnPoint {
     fn show_inspector_ui<'s>(
         &mut self,
         _: &'s Scene,
+        _: &mut hecs::CommandBuffer,
         _: EntityRef<'s>,
         ui: &mut Ui,
         _: &Resources,